refactor: window 生成を window_manager.rs に分離し lib.rs を command 登録に絞る (#14)
