+++ frontend/src-tauri/src/lib.rs
@@ -4,16 +4,15 @@ pub mod pattern_expander;
 mod pattern_loader;
 pub mod patterns;
 pub mod playlist;
+mod window_manager;
 
 use calibration::{CalibrationStatus, ControlResult};
-use displays::{print_display_list, select_displays, DisplayInfo};
 use pattern_loader::load_pattern_with_params;
 use patterns::{load_patterns, PatternsResponse};
 use playlist::Playlist;
 use serde_json::Value;
 use std::collections::HashMap;
 use std::path::PathBuf;
-use tauri::{Manager, WebviewUrl, WebviewWindowBuilder};
 
 /// Tauri command: Get all available patterns
 #[tauri::command]
@@ -81,41 +80,7 @@ pub fn run(pattern: String, display_spec: String, list_displays: bool) {
     tauri::Builder::default()
         .plugin(tauri_plugin_single_instance::init(
             move |app, args, _cwd| {
-                // This callback is triggered when a second instance is launched
-                log::info!(
-                    "Single instance: Another instance attempted to start with args: {args:?}"
-                );
-
-                // If args contain --pattern, switch to that pattern
-                if let Some(pattern_arg_idx) = args.iter().position(|arg| arg == "--pattern") {
-                    if let Some(new_pattern) = args.get(pattern_arg_idx + 1) {
-                        log::info!("Switching to pattern: {new_pattern}");
-
-                        // Switch pattern on all display windows by navigating to new URL
-                        for i in 0..8 {
-                            let window_label = format!("display-{i}");
-                            if let Some(window) = app.get_webview_window(&window_label) {
-                                let url = if cfg!(debug_assertions) {
-                                    format!("http://localhost:3000/?pattern={new_pattern}")
-                                } else {
-                                    format!("tauri://localhost/?pattern={new_pattern}")
-                                };
-                                let js = format!(
-                                    "window.location.href = '{}';",
-                                    url.replace('\'', "\\'")
-                                );
-                                let _ = window.eval(&js);
-                            } else {
-                                break;
-                            }
-                        }
-                    }
-                }
-
-                // Bring existing windows to front
-                if let Some(window) = app.get_webview_window("display-0") {
-                    let _ = window.set_focus();
-                }
+                window_manager::handle_second_instance(app, args);
             },
         ))
         .setup(move |app| {
@@ -127,73 +92,7 @@ pub fn run(pattern: String, display_spec: String, list_displays: bool) {
                 )?;
             }
 
-            // Get available monitors
-            let monitors = app.available_monitors()?;
-            let primary_monitor = app.primary_monitor()?.expect("No primary monitor found");
-
-            // Convert monitors to DisplayInfo
-            let all_displays: Vec<DisplayInfo> = monitors
-                .iter()
-                .enumerate()
-                .map(|(i, m)| {
-                    let is_primary = m.name() == primary_monitor.name();
-                    DisplayInfo::from_monitor(m, i, is_primary)
-                })
-                .collect();
-
-            // If --list-displays, print and exit
-            if list_displays {
-                print_display_list(&all_displays);
-                std::process::exit(0);
-            }
-
-            // Select displays based on specification
-            let selected_displays = select_displays(&display_spec, &all_displays);
-
-            if selected_displays.is_empty() {
-                eprintln!("[ERROR] No displays selected with spec: {display_spec}");
-                std::process::exit(1);
-            }
-
-            // Build URL with pattern parameter
-            let url = if cfg!(debug_assertions) {
-                format!("http://localhost:3000/?pattern={pattern}")
-            } else {
-                format!("tauri://localhost/?pattern={pattern}")
-            };
-
-            // Get the default window (created by tauri.conf.json)
-            // We'll close it and create new ones for each display
-            if let Some(main_window) = app.get_webview_window("main") {
-                main_window.close()?;
-            }
-
-            // Create a window for each selected display
-            for (i, display) in selected_displays.iter().enumerate() {
-                let window_label = format!("display-{i}");
-
-                WebviewWindowBuilder::new(
-                    app,
-                    &window_label,
-                    WebviewUrl::External(url.parse().unwrap()),
-                )
-                .position(display.x as f64, display.y as f64)
-                .inner_size(display.width as f64, display.height as f64)
-                .resizable(false)
-                .fullscreen(false) // We manually set size to display size
-                .decorations(false)
-                .build()?;
-
-                log::info!(
-                    "Created window {} on display {} ({}x{} at {},{})",
-                    window_label,
-                    display.name,
-                    display.width,
-                    display.height,
-                    display.x,
-                    display.y
-                );
-            }
+            window_manager::setup_windows(app, &pattern, &display_spec, list_displays)?;
 
             Ok(())
         })
